participant module incl initializer and signer
