+++ mithril-stm/benches/size_benches.rs
@@ -5,9 +5,8 @@ use blake2::{
 };
 use mithril_stm::key_reg::KeyReg;
 use mithril_stm::stm::{
-    CoreVerifier, Stake, StmClerk, StmInitializer, StmParameters, StmSig, StmSigRegParty,
-    StmSigner, StmVerificationKey,
-};
+    CoreVerifier, Stake, StmClerk, StmParameters, StmSig, StmSigRegParty};
+use mithril_stm::participant::{StmInitializer, StmSigner, StmVerificationKey};
 use rand_chacha::ChaCha20Rng;
 use rand_core::{RngCore, SeedableRng};
 use rayon::iter::ParallelIterator;
