+++ mithril-stm/examples/key_registration.rs
@@ -3,7 +3,8 @@
 //! run presented in `tests/integration.rs`, we explicitly treat each party individually.
 use blake2::{digest::consts::U32, Blake2b};
 use mithril_stm::key_reg::{ClosedKeyReg, KeyReg};
-use mithril_stm::stm::{Stake, StmClerk, StmInitializer, StmParameters, StmVerificationKeyPoP};
+use mithril_stm::stm::{Stake, StmClerk, StmParameters};
+use mithril_stm::participant::{StmInitializer, StmVerificationKeyPoP};
 
 use rand_chacha::ChaCha20Rng;
 use rand_core::{RngCore, SeedableRng};
