+++ mithril-stm/src/merkle_tree/leaf.rs
@@ -1,6 +1,7 @@
 use crate::bls_multi_sig::VerificationKey;
 use crate::error::MerkleTreeError;
-use crate::stm::{Stake, StmVerificationKey};
+use crate::participant::StmVerificationKey;
+use crate::stm::Stake;
 use blake2::Blake2b;
 use digest::consts::U32;
 use serde::{Deserialize, Serialize};
