+++ mithril-stm/src/participant/initializer.rs
@@ -1,0 +1,148 @@
+use crate::bls_multi_sig::{SigningKey, VerificationKeyPoP};
+use crate::key_reg::{ClosedKeyReg, RegParty};
+use crate::participant::StmSigner;
+use crate::stm::{Stake, StmParameters};
+use crate::RegisterError;
+use blake2::digest::Digest;
+use digest::FixedOutput;
+use rand_core::{CryptoRng, RngCore};
+use serde::{Deserialize, Serialize};
+
+/// Wrapper of the MultiSignature Verification key with proof of possession
+pub type StmVerificationKeyPoP = VerificationKeyPoP;
+
+/// Initializer for `StmSigner`.
+/// This is the data that is used during the key registration procedure.
+/// Once the latter is finished, this instance is consumed into an `StmSigner`.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StmInitializer {
+    /// This participant's stake.
+    pub stake: Stake,
+    /// Current protocol instantiation parameters.
+    pub params: StmParameters,
+    /// Secret key.
+    pub(crate) sk: SigningKey,
+    /// Verification (public) key + proof of possession.
+    pub(crate) pk: StmVerificationKeyPoP,
+}
+
+impl StmInitializer {
+    /// Builds an `StmInitializer` that is ready to register with the key registration service.
+    /// This function generates the signing and verification key with a PoP, and initialises the structure.
+    pub fn setup<R: RngCore + CryptoRng>(params: StmParameters, stake: Stake, rng: &mut R) -> Self {
+        let sk = SigningKey::gen(rng);
+        let pk = StmVerificationKeyPoP::from(&sk);
+        Self {
+            stake,
+            params,
+            sk,
+            pk,
+        }
+    }
+
+    /// Extract the verification key.
+    pub fn verification_key(&self) -> StmVerificationKeyPoP {
+        self.pk
+    }
+
+    /// Build the `avk` for the given list of parties.
+    ///
+    /// Note that if this StmInitializer was modified *between* the last call to `register`,
+    /// then the resulting `StmSigner` may not be able to produce valid signatures.
+    ///
+    /// Returns an `StmSigner` specialized to
+    /// * this `StmSigner`'s ID and current stake
+    /// * this `StmSigner`'s parameter valuation
+    /// * the `avk` as built from the current registered parties (according to the registration service)
+    /// * the current total stake (according to the registration service)
+    /// # Error
+    /// This function fails if the initializer is not registered.
+    pub fn new_signer<D: Digest + Clone + FixedOutput>(
+        self,
+        closed_reg: ClosedKeyReg<D>,
+    ) -> Result<StmSigner<D>, RegisterError> {
+        let mut my_index = None;
+        for (i, rp) in closed_reg.reg_parties.iter().enumerate() {
+            if rp.0 == self.pk.vk {
+                my_index = Some(i as u64);
+                break;
+            }
+        }
+        if my_index.is_none() {
+            return Err(RegisterError::UnregisteredInitializer);
+        }
+
+        Ok(StmSigner::set_stm_signer(
+            my_index.unwrap(),
+            self.stake,
+            self.params,
+            self.sk,
+            self.pk.vk,
+            closed_reg,
+        ))
+    }
+
+    /// Creates a new core signer that does not include closed registration.
+    /// Takes `eligible_parties` as a parameter and determines the signer's index in the parties.
+    /// `eligible_parties` is verified and trusted which is only run by a full-node
+    /// that has already verified the parties.
+    pub fn new_core_signer<D: Digest + Clone + FixedOutput>(
+        self,
+        eligible_parties: &[RegParty],
+    ) -> Option<StmSigner<D>> {
+        let mut parties = eligible_parties.to_vec();
+        parties.sort_unstable();
+        let mut my_index = None;
+        for (i, rp) in parties.iter().enumerate() {
+            if rp.0 == self.pk.vk {
+                my_index = Some(i as u64);
+                break;
+            }
+        }
+        if let Some(index) = my_index {
+            Some(StmSigner::set_core_signer(
+                index,
+                self.stake,
+                self.params,
+                self.sk,
+                self.pk.vk,
+            ))
+        } else {
+            None
+        }
+    }
+
+    /// Convert to bytes
+    /// # Layout
+    /// * Stake (u64)
+    /// * Params
+    /// * Secret Key
+    /// * Public key (including PoP)
+    pub fn to_bytes(&self) -> [u8; 256] {
+        let mut out = [0u8; 256];
+        out[..8].copy_from_slice(&self.stake.to_be_bytes());
+        out[8..32].copy_from_slice(&self.params.to_bytes());
+        out[32..64].copy_from_slice(&self.sk.to_bytes());
+        out[64..].copy_from_slice(&self.pk.to_bytes());
+        out
+    }
+
+    /// Convert a slice of bytes to an `StmInitializer`
+    /// # Error
+    /// The function fails if the given string of bytes is not of required size.
+    pub fn from_bytes(bytes: &[u8]) -> Result<StmInitializer, RegisterError> {
+        let mut u64_bytes = [0u8; 8];
+        u64_bytes.copy_from_slice(&bytes[..8]);
+        let stake = u64::from_be_bytes(u64_bytes);
+        let params = StmParameters::from_bytes(&bytes[8..32])?;
+        let sk = SigningKey::from_bytes(&bytes[32..])?;
+        let pk = StmVerificationKeyPoP::from_bytes(&bytes[64..])?;
+
+        Ok(Self {
+            stake,
+            params,
+            sk,
+            pk,
+        })
+    }
+}
