+++ mithril-stm/src/participant/mod.rs
@@ -1,0 +1,7 @@
+//! Mithril-stm participant including StmInitializer and StmSigner
+
+mod initializer;
+mod signer;
+
+pub use initializer::{StmInitializer, StmVerificationKeyPoP};
+pub use signer::{StmSigner, StmVerificationKey};
