+++ mithril-stm/tests/stm_core.rs
@@ -1,8 +1,8 @@
 use blake2::Blake2b;
 use digest::consts::U32;
 use mithril_stm::stm::{
-    CoreVerifier, Stake, StmInitializer, StmParameters, StmSig, StmSigner, StmVerificationKey,
-};
+    CoreVerifier, Stake, StmParameters, StmSig};
+use mithril_stm::participant::{StmSigner, StmInitializer, StmVerificationKey};
 use mithril_stm::CoreVerifierError;
 use rand_chacha::ChaCha20Rng;
 use rand_core::{RngCore, SeedableRng};
